# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: IDE Platform API — Replit-like workspace platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user (first user becomes admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Registration disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /workspaces/{workspaceId}/databases:
    get:
      operationId: listDatabases
      tags: [databases]
      summary: List provisioned databases
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Database"

    post:
      operationId: createDatabase
      tags: [databases]
      summary: Provision a database
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DatabaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Database"
  /workspaces/{workspaceId}/databases/{databaseId}:
    delete:
      operationId: deleteDatabase
      tags: [databases]
      summary: Drop a database
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: databaseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /workspaces/{workspaceId}/databases/{databaseId}/reset-password:
    post:
      operationId: resetDatabasePassword
      tags: [databases]
      summary: Reset database password
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: databaseId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Database"

  # ─── DOMAINS ────────────────────────────────────────────────────────────────
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users (admin only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"

    post:
      operationId: adminCreateUser
      tags: [admin]
      summary: Create a new user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/users/{userId}:
    patch:
      operationId: adminUpdateUser
      tags: [admin]
      summary: Update user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"

    delete:
      operationId: adminDeleteUser
      tags: [admin]
      summary: Delete user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /admin/settings:
    get:
      operationId: adminGetSettings
      tags: [admin]
      summary: Get platform settings (admin only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"

    patch:
      operationId: adminUpdateSettings
      tags: [admin]
      summary: Update platform settings (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlatformSettingsUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformSettings"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]

    # ─── AUTH ────────────────────────────────────────────────────────────────
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 50
        fullName:
          type: ["string", "null"]
        email:
          type: string
        password:
          type: string
          minLength: 8
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, username, email, role, plan, createdAt, onboardingCompleted]
      properties:
        id:
          type: string
        username:
          type: string
        fullName:
          type: ["string", "null"]
        email:
          type: string
        role:
          type: string
          enum: [admin, user]
        plan:
          type: string
          enum: [free, pro, team]
        createdAt:
          type: string
          format: date-time
        onboardingCompleted:
          type: boolean
...
    Database:
      type: object
      required: [id, workspaceId, dbType, dbName, dbUser, dbHost, dbPort, createdAt]
      properties:
        id:
          type: string
        workspaceId:
          type: string
        dbType:
          type: string
          enum: [mysql, postgres]
        dbName:
          type: string
        dbUser:
          type: string
        dbHost:
          type: string
        dbPort:
          type: number
        dbPassword:
          type: ["string", "null"]
        connectionUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    DatabaseInput:
      type: object
      required: [dbType]
      properties:
        dbType:
          type: string
          enum: [mysql, postgres]

    # ─── DOMAIN ──────────────────────────────────────────────────────────────
...
    AdminUser:
      type: object
      required: [id, username, email, role, plan, workspaceCount, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        fullName:
          type: ["string", "null"]
        email:
          type: string
        role:
          type: string
          enum: [admin, user]
        plan:
          type: string
          enum: [free, pro, team]
        workspaceCount:
          type: number
        createdAt:
          type: string
          format: date-time
    AdminUserInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        fullName:
          type: ["string", "null"]
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [admin, user]
        plan:
          type: string
          enum: [free, pro, team]
    AdminUserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [admin, user]
        plan:
          type: string
          enum: [free, pro, team]
        password:
          type: ["string", "null"]
...
    PlatformSettings:
      type: object
      required: [platformName, domain, registrationEnabled, defaultPlan]
      properties:
        platformName:
          type: string
        domain:
          type: string
        registrationEnabled:
          type: boolean
        defaultPlan:
          type: string
          enum: [free, pro, team]
        maxWorkspacesPerUser:
          type: number
        maxRamMb:
          type: number
        maxDiskMb:
          type: number
    PlatformSettingsUpdate:
      type: object
      properties:
        platformName:
          type: string
        domain:
          type: string
        registrationEnabled:
          type: boolean
        defaultPlan:
          type: string
          enum: [free, pro, team]
        maxWorkspacesPerUser:
          type: number
        maxRamMb:
          type: number
        maxDiskMb:
          type: number